+++ GyroData/Domain/TableView/Cell/AnalysisTableViewCell.swift
@@ -82,12 +82,3 @@ final class AnalysisTableViewCell: UITableViewCell {
         analysisTypeLabel.text = cellData[indexPath.row].analysisType
     }
 }
-
-extension Date {
-    func formattedString() -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
-        dateFormatter.timeZone = TimeZone(identifier: "UTC")
-        return dateFormatter.string(from: self)
-    }
-}
